pyproject.toml
[tool.poetry]
name = "quanti-gin"
version = "0.1.0"
description = "A customizable data generator for quantum simulation"
authors = ["Korbinian Stein <kstein@inforbi.de>"]
license = "MIT"
readme = "README.md"
packages = [{include = "quanti_gin"}]

[tool.poetry.dependencies]
python = "3.10.*"
tequila-basic = "^1.9.5"
numpy = "^1.26.4"
pandas = "^2.2.2"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
